requirements.txt
googlesearch-python==1.2.3
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
flask==3.0.0
